Finalizada correção Confirmar Item e mudança de tipo dos campos quantidade para Decimal

+++ server/doacoes/models.py
@@ -11,8 +11,8 @@ class Categoria(models.Model):
 class Demanda(models.Model):
     ong = models.ForeignKey("administrativo.Ong", related_name='demanda', on_delete=models.DO_NOTHING)
     categoria = models.ForeignKey("Categoria", related_name='demanda',on_delete=models.DO_NOTHING)
-    quantidade_solicitada = models.IntegerField("Quantidade solicitada")
-    quantidade_alcancada = models.IntegerField("Quantidade alcançada", blank=True, null=True, default=0)
+    quantidade_solicitada = models.DecimalField("Quantidade solicitada", max_digits=10, decimal_places=2)
+    quantidade_alcancada = models.DecimalField("Quantidade alcançada",max_digits=10, decimal_places=2, blank=True, null=True, default=0)
     unidade = models.CharField("Unidade", max_length=5)
     #data_inicio = models.DateField("Data inicial")
     #data_fim = models.DateField("Data fim")
@@ -28,8 +28,8 @@ class Doacao(models.Model):
     data_confimacao = models.DateField("Data confirmação", blank=True, null=True)
 
 class ItemDoacao(models.Model):
-    quantidade_prometida = models.IntegerField("Quantidade prometida")
-    quantidade_efetivada = models.IntegerField("Quantidade efetivada", blank=True, null=True, default=0)
+    quantidade_prometida = models.DecimalField("Quantidade prometida", max_digits=10, decimal_places=2)
+    quantidade_efetivada = models.DecimalField("Quantidade efetivada", max_digits=10, decimal_places=2, blank=True, null=True, default=0)
     data_atualizacao = models.DateTimeField("Data atualização", auto_now=True)
     doacao = models.ForeignKey("Doacao", related_name='item_doacao', on_delete=models.DO_NOTHING)
     demanda = models.ForeignKey("Demanda", related_name='item_doacao', on_delete=models.DO_NOTHING)
